pyproject.toml
[project]
name = "renderdoc-mcp"
version = "1.0.0"
description = "MCP server for RenderDoc graphics debugging"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastmcp>=2.0.0,<3.0.0",
    "pydantic>=2.0.0",
    "pillow>=9.0.0",  # For image export functionality
    "numpy>=1.20.0",  # For numerical operations
]

[project.scripts]
renderdoc-mcp = "mcp_server.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["mcp_server"]

[tool.uv]
package = true
dev-dependencies = []
